Disable 'follow all' button
Too many people ran into an error when trying this. So now the link is
removed from the template (but the functionality remains in the code),
and the descriptions on the front page and the grab page was changed
accordingly. And as this removes a whole class of errors, these were
all removed from the help page, too.

Since manually adding one account after another is now required, these
are shuffled instead of being sorted on the grab page.